% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{h_nr_simulate}
\alias{h_nr_simulate}
\title{Simulation functions
Simulate numbers of cases with symptoms on day \code{n} that are reported on day \code{r}}
\usage{
h_nr_simulate(num_cases = 100, k = 2, lambda = 9, N = 30, seed = 42)
}
\arguments{
\item{num_cases}{how many cases to simulate}

\item{k}{shape parameter of Weibull delay function}

\item{lambda}{scale parameter of Weibull delay function}

\item{N}{maximum day of reporting, days start at 0}

\item{seed}{seed for random number generator}
}
\value{
h_nr matrix of number of cases, with \code{h_nr[n+1, r+1]} corresponding to
\code{h_nr} (since \code{n} and \code{r} start at 0)
}
\description{
Returns a matrix (that's how I first did it), so modify at some point to
return dataframe.
}
\author{
Andrew Edwards
}
