% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_time_to_report}
\alias{plot_time_to_report}
\title{Plot time from onset of symptoms to case being reported.}
\usage{
plot_time_to_report(
  data,
  start_date_report = "2020-03-10",
  start_date_onset = "2020-03-05",
  show_weekends = FALSE,
  x_axis = "reported",
  xLim = c(lubridate::ymd("2020-02-28"), lubridate::ymd("2020-04-07")),
  yLab = "Time from symptom onset\\n to reported case (days)"
)
}
\arguments{
\item{data}{tibble with columns XX}

\item{start_date_report}{First date of reported case to show.}

\item{start_date_onset}{First date of symptom onset to show}

\item{show_weekends}{Highlight weekends or not}

\item{x_axis}{Either `reported` to show date of case being reported, or
`onset` for date of symptom onset}

\item{xLim}{Limits of dates for x-axis}

\item{yLab}{Label for y-axis}
}
\value{

}
\description{
Bar plots for each day, plus points for each case. Adapted from Mike
 Irvine's code, could replace with Sean's later code that was used in the
 manuscript. Need bit more description.
}
\author{
Andrew Edwards adapted from Mike Irvine's code
}
