% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{h_nr_one_sim_fit}
\alias{h_nr_one_sim_fit}
\title{Wrapper function for simulating one h_nr matrix and fitting with likelihood
using `nlm()`}
\usage{
h_nr_one_sim_fit(init = c(3, 15), ...)
}
\arguments{
\item{init}{initial conditions for optimisation}

\item{...}{inputs for `h_nr_simulate()`}
}
\value{
MLE_res output from nlm, a list containing components:
 - `minimum`: minimum of negative log-likelihod
 - `estimate`: vector of MLEs of `k` and then `lambda`
 - `gradient`: vector of the gradient at the minimum (should be close to `c(0,0)`)
 - `code`: integer indicating why the optimization process terminated, 1 is
  relative gradient is close to zero (see `?nlm`).
 - `iterations`: number of iterations performed
}
\description{
Wrapper function for simulating one h_nr matrix and fitting with likelihood
using `nlm()`
}
\author{
Andrew Edwards
}
