% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_tidy_delay_data}
\alias{load_tidy_delay_data}
\title{load-data.R - loading in and tidying up data
Load in the detailed linelist data and exclude un-needed columns}
\usage{
load_tidy_delay_data()
}
\value{
list of two tibbles with columns \code{reported_date}, \verb{symptom_onset_date and }time_to_report` with each row corresponding to a positive test:
\itemize{
\item delay_data_with_outliers  -- with the outliers still included
\item delay_data                -- with outliers excluded, to use for analyses
}
}
\description{
Load in the detailed linelist data and exclude un-needed columns (that may
want to be looked at some point) and outliers of negative delays and delays30 days (since likely data-input errors). Data are not public (yet) and sothis function will not work for everyone.
}
\author{
Andrew Edwards
}
